CMake: modulemap file override as build interface
We should only be pulling in the modulemap file from the source
directory when compiling against the build directory, not the installed
location. The module-map-file override should only be applied as part of
the build interface, not install interface. In cases where the sources
are checked out and we are compiling a project against the install
interface, we will run into a situation where the compiler sees the
modulemap from both the installed location and the source directory and
will complain.